refactor:채팅방 생성 코드 수정
